+++ esp_cam_sensor/sensors/tc358743/private_include/tc358743_mipi_2lane_27Minput_1920x1080_rgb888_30fps.h
@@ -1,0 +1,8 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+static const tc358743_reginfo_t tc358743_mipi_2lane_27Minput_1920x1080_rgb888_30fps[] = {
+    {0xff, 0xe0},
+};